+++ tenancy/__init__.py
@@ -1,5 +1,0 @@
-"""Tenancy service package."""
-
-from .app import app
-
-__all__ = ["app"]